limit maximum size of image for assi_2 comments
fixes COMMS-2133

The image that displays when no comments have been made on an
attempt in Assignments 2 is too large in the Firefox browser. The
image has been given a maximum size limit so that it remains the
same size in browsers that have no bug, but is limited to an easily
viewable size in Firefox and any other browsers that may have this
same issue.

Test Plan:
* Masquerade as a student in Firefox.
* View a submission attempt in Assignments 2 as a student that has
  no comments made in the "Comments" tab.
* Click on the comments tab; notice that the image of the two
  people is an appropriate size (small enough to view without
  scrolling), and that both the image and the text underneath is
  centered.

Change-Id: I9b1770e5e48f52d2876637cc2b4aa015be606854
Reviewed-on: https://gerrit.instructure.com/198349
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Tested-by: Jenkins
QA-Review: Landon Gilbert-Bland <lbland@instructure.com>
Product-Review: Michelle Simmons <misimmons@instructure.com>