Mobile responsive navigation
closes: COREFE-171

This is meant to be a mergeable first step towards making canvas
responsive to phones. There are things we will still want to do before
we turn it on for real users but this is meant to be mergeable as-is
behind the feature flag. Meaning it does not affect current users in
any way.

Things we know we still want to do:

* find all the context tab icon types and add them to the map
 design to provide icons for 'ePortfolio', 'Permissions', 'Admin Tools',
 'Sub-Accounts' - didn't have anything existing that would work

Test plan:
* check this out
* without the feature flag turned on, nothing should change. To test
  that, you can resize your screen to be small and it canvas should look
  the same as it always has
* turn on the ‘responsive_layout’ feature flag on the domain root
  account
* you should see the mobile responsive view if you resize your screen
  small.

Change-Id: Ie5f3a68a7915cd87b74e549778c1d63f05bc252f
Reviewed-on: https://gerrit.instructure.com/200715
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
Product-Review: Pam Hiett <phiett@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>