Post/hide with separate assignment object
Under very specific circumstances, an assignment posting or hiding
submissions could have a read-only quiz attached to it. Said quiz would
also have modified attributes, with the result that updating the
assignment would also attempt to save the quiz and cause a
ReadOnlyRecord error. When responding to changes in an individual
submission's posted status, we now do the actual posting/hiding on a
separate copy of the assignment (with no readonly quiz) to avoid the
error.

This is a second attempt to solve the issues in the below-mentioned
ticket.

fixes GRADE-2317

Test plan:

(Please follow the directions precisely! There is no room for frivolity
here!)

- Have a course new gradebook and post policies enabled
  - For your sanity, don't have very many students
- Create a quiz with a very short time limit (such as 1 min)
  - Take note of the URL for the quiz; the format will be:
    - /courses/<course ID>/quizzes/<quiz ID>
- As every student in the class but one, take the quiz normally
- As the last student
  - Start taking the quiz
  - You can fill out answers if you like, but DO NOT SUBMIT THE QUIZ.
    Instead, navigate away from the page (e.g. to the course home page)
    before the time limit runs out, ignoring whatever warnings the
    browser gives you.
    - (This will ensure the quiz ends up in a state of "taken, and time
      has run out, but not actually graded")
  - Using a stopwatch or something, determine when the time limit has
    been reached
  - After this point, go directly back to the quiz URL mentioned above
    - This will actually grade the now-completed quiz
    - (Do NOT go to the "Quizzes" index page, as that also seems to
      grade the quiz but via a separate path that did not have the same
      issue)
  - You should see your quiz results, and there should be no error

Change-Id: I3c4b0b46678594576a21d69ee02cdc2885c9a58f
Reviewed-on: https://gerrit.instructure.com/202103
Tested-by: Jenkins
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>
QA-Review: James Butters <jbutters@instructure.com>
Product-Review: KC Naegle <knaegle@instructure.com>