add submission.posted in graphql
With Post Policies enabled, we need to support the
submission.posted attribute in favor of assignment.muted.
When accessing submission.posted, the submission model will
internally fall back to assignment.muted if Post Policies are
not enabled, so this change should work whether the feature is
enabled or not.

closes COMMS-2217

TEST PLAN:
in a course with New Gradebook and Post Policies enabled
- an assignment has its post policy set to manually post grades
- a student submits to that assignment
- a grade and comment are given for that submission
- graphql shows submission.posted == false for that submission
-- this can be check on the /graphiql page or by checking the network call
   made to graphql when the student visits the A2 page for the assignment
   ( assignment > submissionsConnection > nodes > 0 > posted )
- the student cannot see the comment given in the Comments tab
- the grade is manually posted by the instructor
- graphql now shows submission.posted == true
- the student can see their grade and the comment given

Change-Id: I2f9a8df18139bcd0dbaf77433b07246d39cbc19e
Reviewed-on: https://gerrit.instructure.com/202472
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Tested-by: Jenkins
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>