Use thumbnailed images from inst-fs for dashcards
Closes: COREFE-295

This will add some extra query string parameters to the url we send
to InstFS so it knows to resize the image down to the dimensions of
a dashcard (or 2x that if the user comes from a high DPI screen).

All it does is add the query string parameters so even if it is ran
in an environment that doesn’t use instFS, it will still work, it just
won’t actually resize the image down

Note: we add the query string to ask for a webp format here but that
functionality hasn’t been added yet to InstFS at the time of writing
this. Once it does have that it will serve a webp image when it is
smaller but for now it will just serve the image in whatever format 
it was uploaded in.

Test plan:
* to test this locally, just verify that the new query string parameters
  Get added and that everything else still works
* once this is on beta or if you test in an environment that does
  use instFS, verify that instFS serves the smaller image


Change-Id: If095c9f1db6f0c07585dd2a15f6252c7351599f0
Reviewed-on: https://gerrit.instructure.com/209155
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Reviewed-by: Jonathan Featherstone <jfeatherstone@instructure.com>
Tested-by: Jenkins
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>