refactor alerts in direct share course tray
This is a preparatory commit for reusing the tray for a content share
import.

It also includes a minor change in behavior to the tray: when the copy
operation has started, the "copy" and "cancel" buttons are replaced with
a single "close" button.

The mechanism for doing these particular screen reader alerts has also
changed to something that should be more reliable. Instead of using the
static alert region on the page, this simply adds the screen reader
alert properties to the alert text. Modern screen readers seem to handle
this better than the static alert div.

refs ADMIN-2912
flag=direct_share

test plan:
- should still be able to copy a discussion from one course to another
  as before, with the changes to the button behavior specified above.
- screen reader alerts for the operation status should still work.

Change-Id: I46395b0c70cd111f455feac7f7d8d44b785552d1
Reviewed-on: https://gerrit.instructure.com/211843
Tested-by: Jenkins
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>