Fix buggy link to external tool in user navigation
If windowTarget=blank, the link appears as something like:

/users/1234/external_tools/123.display=borderless

Instead of

/users/1234/external_tools/123?display=borderless

This URL generation from the tab API can also generate full URLs --
with a different host -- and the way we do that was messing up
query parameters since query parameters are sent as a hash as the last
argument in `args` for the tab. (The other code path which renders the
tabs directly, e.g. in the account navigation, doesn't support a full
URL with a different host so doesn't have this issue)

This change could potentially affect other links generated by the tab
API, however it's highly unlikely we ever really wanted to pass in a hash
(which normally gets serialized as `key1=val1&key2=val2...`) to the rails
URL helper as anything but the last parameter. The other tabs which
could have an `args` that I found were:

1. Under account navigation,  `lti_apps_tabs` in
  `models/lti/message_handler.rb` can generate tabs with
  `args = {message_handler_id: app.id, resource_link_fragment: "nav"}`.
2. In a Course, `tab_configuration` can apparently come from
   the database so it could have something arbitrary.

I talked thru these cases with Weston and he thinks it's unlikely that
these would be negatively affected by this change.

Closes PLAT-4800

Test plan
- install an LTI tool with placement in the user navigation that uses
  `windowTarget` = `_blank`. The link in the user navigation should
  end with `?display=borderless`. (Note that it seems that we don't
  actually support a borderless launch from the user navigation -- when
  the link is clicked, it still has the canvas navigation)
- try to find some of the other cases that could possibly be affected
  by this change (?)

Change-Id: I27cc1a3da37c4d7e4da4fabefc38c222b3eaa73d
Reviewed-on: https://gerrit.instructure.com/212079
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Tested-by: Jenkins
Product-Review: Jesse Poulos <jpoulos@instructure.com>