Use 1.year for cache ttl for course.image
This should make it so we get better caching for the dashcard image
urls.

Right now, the ttl is the default of 1.day like everything else. But
Since course.image is used so frequently and is not sensitive
like things like a submission attachment or something we can safely
tell instFS to cache it for up to a year

Change-Id: I14123c52eef799c4dc901bd5a39a664c7a2c1d03
Reviewed-on: https://gerrit.instructure.com/213241
Tested-by: Jenkins
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Reviewed-by: Jonathan Featherstone <jfeatherstone@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>