Bring js_env :direct_share flag to application_controller
Closes ADMIN-3005

The left nav code on the front end will need to check the
state of the :direct_share feature flag. Since the nav is
shown everywhere this rates being in the application_controller
so we will hoist it there and then remove it from the ones
further down the hierarchy.

In the quizzes_controller, the setting of the FF is also
unfortunately dependent on a permission, so we're going to
leave that one alone for now, and fix it later.

Test plan:
* Tests all still pass
* Not sure? What UI features are sensitive to that FF?

Change-Id: Id2db845a1c35cce08bf2cb99b5893ae23f70e045
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/213894
Reviewed-by: Rex Fleischer <rfleischer@instructure.com>
Tested-by: Jenkins
QA-Review: Rex Fleischer <rfleischer@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>