Add OriginalityReports API #create `attempt` param
Add `attempt` parameter to the #create (POST) endpoint of the originality
reports API.

The #create endpoint can be used to create or update an originality
report. When there is a file ID, we use that to find an existing originality
report with that file ID to update it. However, for assignments that
allow online text entries, submissions (and thus originality reports)
may not have a file ID. So we have no way to know if the client (e.g.,
Turnitin) meant to create a new originality report, or update an
existing one.

This change adds an "attempt" parameter, so we can find the correct
version of the submission. If creating a new report, it also sets
the new report's submission_time to match the submission version's
submitted_at, so the UI knows which report corresponds to which version.

To fully fix the issue in the ticket, we will require plagiarism
detection tools like Turnitin to use this new parameter.

Test plan (sorry):
1. Create an assignment that allows Online Text Entry submissions.
2. Set up lti_originality_report_example and use it with the assignment.
   See https://github.com/instructure/lti_originality_report_example
3. As a student (e.g. /.../assignment/123?become_student) make a submission.
4. Get the id for that submission from the Canvas DB:
   `Submission.last.id`
5. In lti_originality_report_example, in app/helpers/submissions_helper.rb,
   add a `puts access_token` in `retrieve_submission`. Then, go to the
   Similarity Detection Tool in the Canvas course and click "Fetch
   submission" on any submission. This will give us an access token
   to use for API calls.
6. Create an originality report for the submission using the API's
   POST (create) endpoint for attempt 1, e.g.:
   ```
   curl -H 'Authorization: Bearer access_token_from_lti_originality_report_example' \
     http://web.canvas-lms.docker/api/lti/assignments/ASSIGNMENT_ID/submissions/SUBMISSION_ID/originality_report \
     --data 'originality_report[attempt]=1&originality_report[originality_score]=55'
   ```
7. Make another submission.
8. Create another originality report for that submission for
   attempt 2, with a different score.
9. Use SpeedGrader on the assignment in Canvas to see that
   two originality reports were created, and that the
   scores match up correctly.
10. Update one of the scores using the POST (create) endpoint
   and in the database, make sure no additional
   OriginalityReport objects were created.
11. If you want, you could test other scenarios by creating both
   originality reports in either order, after making both submissions,
   no attempt given, etc.

Closes PLAT-4817

Change-Id: Idfa9fdc698cbf0b7353cfdea4e0c80e46e4e93c7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/214765
Tested-by: Evan Battaglia <ebattaglia@instructure.com>
Reviewed-by: Marc Phillips <mphillips@instructure.com>
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Product-Review: Jesse Poulos <jpoulos@instructure.com>