force reply-to-address to less than 64 chars
the reply to address needs to have less than 64 chars before the @. the
first characters are presumably `reply+`, but can also be set to
`notifications+`. After this we create a secure id that is now truncated
to 16 characters, followed by the shortened global id of the message and
the timestamp to_i.

truncating the secure id allows us to guarantee that the address will
not exceed 64 chars and we can still use the same processing logic for
older messages.

test plan
 - generate a message
 - apply change
 - generate a new message
 - the reply to address should be shorter than it was before
 - both replies should work and process

fixes WOKE-87

Change-Id: If82dd61ec5b29aa079c82043e206b0094f0e4ae9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/215844
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>