Move test suite builds to new folder
This adds the folder to the path for each test suite job.
Merging this will break the builds until I move the jobs in
Jenkins, so I plan to merge this in the evening and immediately
move the jobs.

Fixes: CCI-12

Test plan:
This patchset won't pass the New Jenkins jobs until their moved, so
we'll have to CR on faith. If there is a problem, I'll revert.

Change-Id: I9d214eab5cc4fa1df636ce31be6a7d39181cde46
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/216060
Tested-by: Jenkins
Reviewed-by: Gentry Beckmann <gbeckmann@instructure.com>
QA-Review: Robert Lamb <rlamb@instructure.com>
Product-Review: Robert Lamb <rlamb@instructure.com>