Use id instead of label to match on a profile tab
Fixes ADMIN-3044
flag=direct_share

It was pointed out that the matching of which count to
associate with which profile tab was being done by the
label, but the label is a translated string so it would
only work in English. So now we match on the tab id
instead. Not sure why I didn't do that the first time.

Test plan:
* Have a user with an unread direct share
* Their unread badges on the left nav and on the
  profile tray should still work just as before

Change-Id: Iae61cb3c06570564e525908b92ec04543fdfd9d3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/216202
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Jenkins
QA-Review: Steve Shepherd <sshepherd@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>